Push error handling down into concrete implementations
